botSolver debug adjustment + Main +add decorator
