+++ Package/__init__.py
@@ -1,1 +1,1 @@
-__all__ = [ 'analyseBot' , 'screenBot'  , 'solverBot' ] 
+__all__ = [ 'analyseBot' , 'screenBot'  , 'solverBot' , 'decoratori' ]
