+++ Package/analyseBot.py
@@ -3,6 +3,7 @@ import numpy as np
 import sys
 import time
 from Package import solverBot
+from Package import decoratori
 
 
 #GLOBALS
